added validations/refactor

+++ sam2/src/main/java/com/wellsfargo/sam2/controllers/EmployeeController.java
@@ -69,11 +69,17 @@ public class EmployeeController {
 //    }
 
     @PostMapping("/add")
-    public ResponseEntity<EmployeeMaster> createEmployee(@RequestBody EmployeeMaster employee) {
+    public ResponseEntity<?> createEmployee(@RequestBody EmployeeMaster employee) {
         try {
             String iempId = employee.getEmployeeId();
-            EmployeeMaster newEmployee = employeeRepository.save(employee);
-            return new ResponseEntity<>(newEmployee, HttpStatus.CREATED);
+            if(employeeRepository.existsById(iempId))
+            {
+                return new ResponseEntity<>(new CustomResponse("cannot add: already in database","failed"), HttpStatus.INTERNAL_SERVER_ERROR);
+            }
+            else {
+                EmployeeMaster newEmployee = employeeRepository.save(employee);
+                return new ResponseEntity<>(newEmployee, HttpStatus.CREATED);
+            }
         } catch (Exception e) {
         	System.out.println("WE have Employeemaster ex" + e);
             return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
@@ -106,10 +112,20 @@ public class EmployeeController {
     @PostMapping("/delete/{id}")
     public ResponseEntity<CustomResponse> deleteEmployee(@PathVariable String id) {
         try {
-        	empCardDetSerImp.deleteByEmployeeId(id);
-        	issueDetailsService.DeleteByEmployeeId(id);
-            employeeRepository.deleteById(id);
-            return new ResponseEntity<>(new CustomResponse("Employee deleted successfully!","Success"),HttpStatus.NO_CONTENT);
+            if(employeeRepository.existsById(id)) {
+
+                if(empCardDetSerImp.existEmployeeCardDetailsById(id)) {
+                    empCardDetSerImp.deleteByEmployeeId(id);
+                }
+                if(issueDetailsService.existsIssueDetailsById(id)) {
+                    issueDetailsService.DeleteByEmployeeId(id);
+                }
+                employeeRepository.deleteById(id);
+                return new ResponseEntity<>(new CustomResponse("Employee deleted successfully!", "Success"), HttpStatus.NO_CONTENT);
+            }
+            else {
+                return new ResponseEntity<>(new CustomResponse("Not found in database","failed"),HttpStatus.INTERNAL_SERVER_ERROR);
+            }
         } catch (Exception e) {
         	System.out.println("Error in delete "+ e);
             return new ResponseEntity<>(new CustomResponse("Some error in server!","failed"),HttpStatus.INTERNAL_SERVER_ERROR);
